requirements.txt
PyQt6>=6.9.0
numpy>=2.2.0
pillow>=11.0.0 